% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_by_drug_name}
\alias{class_by_drug_name}
\title{Get the classes containing a specifed drug name as a member.}
\usage{
class_by_drug_name(drugName, relaSource = NULL, relas = NULL,
  trunc = T)
}
\description{
The user may specify a source of drug-class relations and relationship values
to filter the output returned.
}
\examples{
class_by_drug_name(drugName = 'armodafinil')
class_by_drug_name(drugName = 'armodafinil', relaSource = 'MEDRT')
class_by_drug_name(drugName = 'armodafinil', relaSource = 'MEDRT', relas = c('has_MOA', 'has_PE'))
}
