% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_by_id}
\alias{class_by_id}
\title{Retrieve class information from a class identifier.}
\usage{
class_by_id(classId, trunc = T)
}
\description{
Intakes a classId and outputs the class name and class type
}
\examples{
class_by_id(classId = "B01AA")
class_by_id("D000998", trunc = FALSE)
}
