% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_by_name}
\alias{class_by_name}
\title{Retrieve class information from a class name.}
\usage{
class_by_name(className, classTypes = NULL, trunc = T)
}
\arguments{
\item{className}{The name of the class}

\item{classTypes}{(optional) a list of class types to check. See
/classtypes for a list of the possible class types. If none are specified,
all class types are checked for the name.}
}
\description{
Retrieve class information from a class name.
}
\examples{
class_by_name(className = "radiopharmaceuticals")
class_by_name(className = "Alkylating Agents", classTypes = 'CHEM')
class_by_name(className = "Alkylating Agents", classTypes = c('CHEM', 'MESHPA'))
}
