% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_by_rxcui}
\alias{class_by_rxcui}
\title{Returns the class type and value of a specified drug}
\usage{
class_by_rxcui(rxcui, relaSource = NULL, relas = NULL, trunc = T)
}
\value{
Dataframe of all the classes for the given drug RxCUI, database(s) and relationship
}
\description{
Intakes a RxCUI identifier and outputs all the classes for the specified
class type and database
}
\examples{
class_by_rxcui(rxcui = 7052, relaSource = "DAILYMED")
class_by_rxcui(rxcui = 7052, relaSource = "MEDRT", relas = "may_treat")
class_by_rxcui(rxcui = 7052, relaSource = "MEDRT", relas = c("may_treat", "ci_with"))
}
