% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_context}
\alias{class_context}
\title{Get the class contexts}
\usage{
class_context(classId, trunc = T)
}
\description{
A context is a path from the root of the class hierarchy down to the specified class.
}
\examples{
class_context('D019275')

}
