% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_graph}
\alias{class_graph}
\title{Get the graph of a specified class.}
\usage{
class_graph(classId, source, trunc = T)
}
\arguments{
\item{classId}{The class identifier}

\item{source}{The class types. See class_types() function for list of all
possible class types}
}
\description{
Returns the ancestors of a specified class in a graph form
}
\examples{
class_graph(classId = 'D003879', source = 'MESHPA')
}
