% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_members}
\alias{class_members}
\title{Get the drug members of a specified class.}
\usage{
class_members(classId, relaSource = NULL, rela = NULL, trans = NULL,
  ttys = NULL, trunc = T)
}
\arguments{
\item{classId}{the class identifier. Note that this is NOT an RxNorm
identifier, but an identifier from the source vocabulary}

\item{relaSource}{the source asserting the relationships between the drug
members and the drug class.}

\item{trans}{(optional) 0 = include indirect and direct relations (the
default). 1 = direct relations only.}

\item{ttys}{(optional) include only drugs with the specified RxNorm term
types. Default is IN, PIN and MIN.}

\item{relas}{the relationship of the drug class to its members. See table
below for valid relationship values.}
}
\description{
This function requires the specification of a source of drug-class
relationships (for example: DailyMed) as well as a relationship to the class.
See the table in the input section for the list of possible values for these
parameters.
}
\examples{
class_members(classId = 'N0000008638', relaSource = 'MEDRT', rela = 'has_PE', ttys = 'IN')
class_members(classId = 'N0000008638', relaSource = 'MEDRT', rela = 'has_PE', ttys = c('IN', 'PIN'))
}
