% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_relas}
\alias{class_relas}
\title{Get the relationships for a source of drug relations.}
\usage{
class_relas(relaSource, trunc = T)
}
\description{
Get the relationships for a source of drug relations.
}
\examples{
class_relas('DAILYMED')
class_relas('medrt')

}
