% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_relaSources}
\alias{class_relaSources}
\title{Get the list of sources that associate generic drugs to the class types.}
\usage{
class_relaSources()
}
\description{
Get the list of sources that associate generic drugs to the class types.
}
\examples{
class_relaSources()
}
