% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_similar}
\alias{class_similar}
\title{Find similar classes of drug members}
\usage{
class_similar(classId, relaSource = NULL, rela = NULL, scoreType = 0,
  top = 10, equivalenceThreshold = NULL, inclusionThreshold = NULL,
  trunc = T)
}
\description{
The input drug class is defined by the class identifier (classId), the source
of drug relations (relaSource) and the relationship (rela) of the drugs to
the class.
}
\examples{
class_similar(classId ='L01XE', relaSource='ATC', scoreType=1, top=4)
class_similar(classId ='D002318 ', relaSource='MEDRT', rela = 'CI_with')
}
