% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_similar_info}
\alias{class_similar_info}
\title{Return the similarity information between two drug classes.}
\usage{
class_similar_info(classId1, relaSource1 = NULL, rela1 = NULL,
  classId2, relaSource2 = NULL, rela2 = NULL, trunc = T)
}
\description{
Return the similarity information between two drug classes.
}
\examples{
class_similar_info(classId1 = 'N02AA', relaSource1 = 'ATC', classId2 = 'D009294', relaSource2 = 'MESH')
}
