% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_tree}
\alias{class_tree}
\title{Get the class tree for a specified class identifier.}
\usage{
class_tree(classId, trunc = T)
}
\description{
The class tree represents the classes which are descendents of the specified class.
}
\examples{
class_tree('D019275')

}
