% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{class_types}
\alias{class_types}
\title{Get the class types.}
\usage{
class_types()
}
\description{
The resources /class/byName and /allClasses use the class types as filters for the output.
}
\examples{
class_types()
}
