% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{classes_all}
\alias{classes_all}
\title{Get all classes for each specified class type.}
\usage{
classes_all(classTypes, trunc = T)
}
\description{
Get all classes for each specified class type.
}
\examples{
classes_all('MOA')

}
