% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbyname.R
\name{getbyname}
\alias{getbyname}
\title{rough approach to dealing with names for rx without functioning RXCUI}
\usage{
getbyname(name = "lovastatin", filter = TRUE)
}
\arguments{
\item{name}{a string}
}
\description{
rough approach to dealing with names for rx without functioning RXCUI
}
