% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcodes.R
\name{getclass}
\alias{getclass}
\title{given rxcui, produce a list from rxclass 'byRxcui'}
\usage{
getclass(num, filter = TRUE)
}
\arguments{
\item{num}{a numeric identifier for a drug}

\item{filter}{logical(1), if TRUE, will extract only the elements with 'minConcept$rxcui' matching `num`}
}
\description{
given rxcui, produce a list from rxclass 'byRxcui'
}
\examples{
chk = getclass(1364447)
length(chk)
}
