% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpath.R
\name{getpath}
\alias{getpath}
\title{decode a classid and provide information on the hierarchy in which it is included}
\usage{
getpath(classid = "B01AF")
}
\arguments{
\item{classid}{character(1) defaults to an ATC1-4 code B01AF}
}
\description{
decode a classid and provide information on the hierarchy in which it is included
}
\examples{
getpath()
}
