% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_func.R
\name{make_upset}
\alias{make_upset}
\title{produce an upset plot for drug categories based
on output of `summ` applied to multiple prescription
series}
\usage{
make_upset(drugdf, nsets_in = 7, nintersects_in = 8, maxlablength = 32)
}
\arguments{
\item{drugdf}{a data.frame as produced by `summ`}

\item{nsets_in}{passed to nsets of upset()}

\item{nintersects_in}{passed to nsets of upset()}

\item{maxlablength}{numeric default to 22, class labels longer than this will be presented via substr}
}
\description{
produce an upset plot for drug categories based
on output of `summ` applied to multiple prescription
series
}
\examples{
data(allcls)
thes = lapply(allcls, summ, 3)
for (i in seq_len(length(allcls))) thes[[i]]$id = i
ddf = do.call(rbind, thes)
make_upset(ddf)
}
