% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{rx_approximate_term}
\alias{rx_approximate_term}
\title{Get RxNorm Approximate Terms}
\usage{
rx_approximate_term(term, maxEntries = 20, option = 0, trunc = T,
  max_rank = 1, min_match_score = 0)
}
\arguments{
\item{term}{Search string.}

\item{maxEntries}{The maximum number of entries to return (default 20)}

\item{option}{Special processing options.  0 - no special processing. 1 -
return only information for terms contained in valid RxNorm concepts. That
is, the term must either be from the RxNorm vocabulary or a synonym of a
(non-suppressed) term in the RxNorm vocabulary.}
}
\value{
Approximate terms.
}
\description{
Do an approximate match search to determine the strings in the data set that
most closely match the search string. See
\href{https://rxnav.nlm.nih.gov/RxNormAPIs.html#uLink=RxNorm_REST_getApproximateMatch}{RxNorm}.
}
\examples{
rx_approximate_term("lipitor")
rx_approximate_term(term = "zocor 2010 20mm", trunc = F)
rx_approximate_term("lipitor", max_rank = 6)
rx_approximate_term("lipitor", max_rank = 6, min_match_score = 40)
}
