% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setsumm.R
\name{setsumm}
\alias{setsumm}
\title{provide a summary for a set of Rx codes that
might represent a patient's profile}
\usage{
setsumm(codes, level = 2, sleeptime = 1)
}
\description{
provide a summary for a set of Rx codes that
might represent a patient's profile
}
\examples{
data(costore)
ss = setsumm(costore[[1]])
}
