% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{summ}
\alias{summ}
\title{produce a summary from a getclass collection}
\usage{
summ(cl, level = 4)
}
\arguments{
\item{cl}{a list of outputs of getclass}

\item{level}{depth in hierarchy of rxClass path}
}
\description{
produce a summary from a getclass collection
}
\examples{
data(allcls)
summ(allcls)
}
