% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getclass.R
\name{getclass}
\alias{getclass}
\title{given rxcui, produce a list from rxclass 'byRxcui'}
\usage{
getclass(num, filter = TRUE, atcOnly = TRUE, addPaths = TRUE)
}
\arguments{
\item{num}{a numeric identifier for a drug}

\item{filter}{logical(1), if TRUE, will extract only the elements with 'minConcept$rxcui' matching `num`}

\item{atcOnly}{logical(1), if TRUE (default) will confine attention to class in ATC1-4, for producing paths}
}
\value{
a list with elements input (the identifier),
items (a matrix with three rows: classId, className, classType), and paths
(output of getpath for the discovered classId, describing path upward through hierarchy)
}
\description{
given rxcui, produce a list from rxclass 'byRxcui'
}
\examples{
chk = getclass(1364447)
str(chk)
}
